COM-2520: create user to add to course
