+++ src/modules/client/pages/ni/courses/LearnersDirectory.vue
@@ -32,6 +32,7 @@ import LearnerCreationModal from '@components/courses/LearnerCreationModal';
 import { NotifyPositive, NotifyNegative } from '@components/popup/notify';
 import { userMixin } from '@mixins/userMixin';
 import { learnerDirectoryMixin } from '@mixins/learnerDirectoryMixin';
+import { learnerCreationMixin } from '@mixins/learnerCreationMixin';
 
 export default {
   metaInfo: { title: 'Répertoire apprenants' },
@@ -41,7 +42,7 @@ export default {
     'ni-table-list': TableList,
     'learner-creation-modal': LearnerCreationModal,
   },
-  mixins: [userMixin, learnerDirectoryMixin],
+  mixins: [userMixin, learnerDirectoryMixin, learnerCreationMixin],
   async created () {
     await this.getLearnerList(this.company._id);
   },
